requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
drf-yasg==1.10.0
Faker==0.9.0
flex==6.13.2
future==0.16.0
idna==2.7
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
pylint==2.1.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
rfc3987==1.3.8
ruamel.yaml==0.15.64
simplejson==3.16.0
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.0
text-unidecode==1.2
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.23
validate-email==1.3
wrapt==1.10.11